[change] change int size from 8 bytes to 4 bytes and add long type

+++ test/include_test.c
@@ -2,7 +2,7 @@
 
 int i;
 
-long long func()
+int func()
 {
     printf("OK!: %d\n", i++);
     return 10;
